# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/Hydraulic/FluidHP/{pressurePsi}/{pumpFlowGpm}": {
      "get": {
        "tags": [
          "Hydraulic"
        ],
        "description": "",
        "parameters": [
          {
            "name": "pressurePsi",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pumpFlowGpm",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Hydraulic/MotorTorqueToInLbs/{pressurePsi}/{motorDisplacementCuInPerRev}": {
      "get": {
        "tags": [
          "Hydraulic"
        ],
        "description": "",
        "parameters": [
          {
            "name": "pressurePsi",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "motorDisplacementCuInPerRev",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Hydraulic/MotorHP/{torqueInLbs}/{motorRpm}": {
      "get": {
        "tags": [
          "Hydraulic"
        ],
        "description": "",
        "parameters": [
          {
            "name": "torqueInLbs",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "motorRpm",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Hydraulic/MotorRPM/{pumpFlowGpm}/{motorDisplacementCuInPerRev}": {
      "get": {
        "tags": [
          "Hydraulic"
        ],
        "description": "",
        "parameters": [
          {
            "name": "pumpFlowGpm",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "motorDisplacementCuInPerRev",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Length/MillimeterToInch/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/CentimeterToInch/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/FootToMillimeter/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/InchToCentimeter/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/InchToMillimeter/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/KilometerToMile/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/MeterToFoot/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/MeterToInch/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/MileToKilometer/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/MicrometerToInch/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/YardToMillimeter/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/RackUToInch/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/InchToRackU/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/InchToHand/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Length/HandToInch/{input}": {
      "get": {
        "tags": [
          "Length"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Rate/FuelCostForRuntime/{numberHoursToRun}/{fuelPricePerUnit}/{burnRateUPH}": {
      "get": {
        "tags": [
          "Rate"
        ],
        "description": "Calculate fuel cost for the desired runtime",
        "parameters": [
          {
            "name": "numberHoursToRun",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Number of hours desired to run"
          },
          {
            "name": "fuelPricePerUnit",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Price of fuel per unit whether gallon or liter, but the unit used here  is the same unit used in the burnRateUPH"
          },
          {
            "name": "burnRateUPH",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Fuel burn rate in your preferred units per hour(gallons, liters, etc.)  as long as the units matches the fuelPricePerUnit"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Rate/FuelTankRuntime/{burnRateUPH}/{fuelTankCapacity}": {
      "get": {
        "tags": [
          "Rate"
        ],
        "description": "Calculate runtime for the fuel tank",
        "parameters": [
          {
            "name": "burnRateUPH",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Fuel burn rate in your preferred units per hour(gallons, liters, etc.)  as long as the units matches the fuelTankCapacity"
          },
          {
            "name": "fuelTankCapacity",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Fuel tank capacity in a unit matching the burnRateUPH"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}